pyproject.toml
[project]
name = "simbirsoft"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Sergey Lazarev", email = "lasest1999@gmail.com" }
]
dependencies = [
    "selenium>=4.21.0",
    "python-dotenv>=1.0.1",
    "pytest>=8.2.0",
    "freezegun>=1.5.1",
    "pyfakefs>=5.5.0",
]
readme = "README.md"
requires-python = ">= 3.8"

[project.scripts]
"simbirsoft" = "simbirsoft:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "mypy>=1.10.0",
    "ruff>=0.4.4",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/simbirsoft"]
